superIdx==-1 means facet in all structures replaced faulty alreadyMoved malloc with vector (MoveSelectedFacets) fixed cloneselectedfacets (not updating sh.nbvertex)

+++ Buffer_shared.h
@@ -114,7 +114,7 @@ public:
 	double area;           // Facet area (m^2)
 
 	int    profileType;    // Profile type
-	size_t    superIdx;       // Super structure index (Indexed from 0)
+	int    superIdx;       // Super structure index (Indexed from 0) -1: facet belongs to all structures (typically counter facets)
 	size_t    superDest;      // Super structure destination index (Indexed from 1, 0=>current)
 	int	 teleportDest;   // Teleport destination facet id (for periodic boundary condition) (Indexed from 1, 0=>none, -1=>teleport to where it came from)
